#105: combine HP Redistribution and Elemental Research into Imp Contract Faire - increase HP Redistribution values, but clamp to delver's current hp
