+++ source/rooms/_roomDictionary.js
@@ -13,10 +13,9 @@ for (const file of [
 	"battle-tortoises.js",
 	"empty.js",
 	"event-artifactdupe.js",
-	"event-elementswap.js",
 	"event-freegoldonfire.js",
 	"event-freerepairkit.js",
-	"event-hpshare.js",
+	"event-impcontractfaire.js",
 	"event-scorebeggar.js",
 	"finalbattle-elkemist.js",
 	"finalbattle-mechaqueen.js",
